requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.8.0
cryptography==42.0.5
fastapi==0.110.0
filelock==3.13.3
firebase-admin==6.5.0
frozenlist==1.4.1
fsspec==2024.3.1
gcloud==0.18.3
google-api-core==2.18.0
google-api-python-client==2.123.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
Jinja2==3.1.3
jwcrypto==1.5.6
lightning-utilities==0.11.2
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauth2client==4.1.3
packaging==24.0
pg8000==1.30.5
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==4.1.0
python-multipart==0.0.9
pytorch-lightning==2.2.1
PyYAML==6.0.1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
scramp==1.4.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlmodel==0.0.16
starlette==0.36.3
sympy==1.12
torch==2.2.2
torchmetrics==1.3.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.29.0
yarl==1.9.4
